Refactored Changes to remove lightcouch wrapping
Removed light couch `Changes` and `ChangesResult`.
Moved method implementations into cloudant package `Changes` and
`ChangesResult` classes.
Refactored constructor and called it from `Database.changes()`
Updated packages in findbugs exclude file.

Added documentation for Changes parameters.

Fixed doc on continuousChanges.
